requirements.txt
flask==3.0.2
flask-cors==4.0.0
PyJWT==2.8.0 